chore: remove duplicate contexts directory (Task 29)
- Remove duplicate contexts from src/contexts/ directory
- Consolidate all context imports to use src/shared/contexts/
- Update all import statements throughout codebase
- Fix import syntax issues and ensure proper module resolution
- Maintain functionality while removing ~600 lines of duplicate code
- Update documentation to reflect new context structure

This cleanup improves maintainability by establishing consistent
context patterns and reducing code duplication.
