+++ docs/architecture/state-management.md
@@ -93,7 +93,7 @@ graph LR
 ### Component Integration
 
 ```jsx
-import { useAppState } from '../../contexts/app';
+import { useAppState } from '../../shared/contexts/app';
 import { useURLSync } from '../../hooks/useURLSync.js';
 
 function EventsDetail() {
@@ -179,7 +179,7 @@ When updating components to use the new state management:
 
 1. **Add Context Hooks**:
 ```jsx
-import { useAppState } from '../../contexts/app';
+import { useAppState } from '../../shared/contexts/app';
 import { useURLSync } from '../../hooks/useURLSync.js';
 ```
