+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "vikings-eventmgmt-mobile",
   "private": true,
-  "version": "2.1.0",
+  "version": "2.3.1",
   "description": "Mobile React version of Vikings Event Management",
   "type": "module",
   "engines": {
