--- src/contexts/app/index.ts
@@ -1,10 +1,0 @@
-export {
-  AppStateProvider,
-  useAppState,
-  AppStateContext,
-  type AppState,
-  type NavigationData,
-  type AppStateAction,
-  type AppStateContextType,
-  type AppStateProviderProps,
-} from './AppStateContext.tsx';
