--- src/contexts/notifications/index.ts
@@ -1,4 +1,0 @@
-export { NotificationProvider, useNotification } from './NotificationContext';
-export type { Notification, NotificationContextType } from './types';
-export { NOTIFICATION_DEFAULTS, createNotificationConfig } from './config';
-export { useNotificationUtils } from './notificationUtils';
