+++ src/features/admin/components/DataClearPage.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { Card } from '../../../shared/components/ui';
-import { useAppState } from '../../../contexts/app';
+import { useAppState } from '../../../shared/contexts/app';
 import databaseService from '../../../shared/services/storage/database.js';
 import { clearToken } from '../../../shared/services/auth/tokenService.js';
 import logger, { LOG_CATEGORIES } from '../../../shared/services/utils/logger.js';
