+++ src/routes/AppRouter.jsx
@@ -1,6 +1,6 @@
 import React, { Suspense } from 'react';
 import { BrowserRouter, Routes, Route, Navigate } from 'react-router-dom';
-import { AppStateProvider } from '../contexts/app';
+import { AppStateProvider } from '../shared/contexts/app';
 import { NotificationProvider } from '../shared/contexts/notifications';
 import LoadingScreen from '../shared/components/LoadingScreen.jsx';
 import ResponsiveLayout from '../shared/components/layout/ResponsiveLayout.jsx';
